added ed's updated scripts

+++ Assets/Scripts/Map Scripts/IncreasePopularityorMoneyScript.cs
@@ -4,8 +4,8 @@ using UnityEngine.UI;
 
 public class IncreasePopularityorMoneyScript : MonoBehaviour {
 	public Button btn;
-	public int amounttoIncreasePop;
-	public int amounttoIncreaseMoney;
+	public float amounttoIncreasePop;
+	public float amounttoIncreaseMoney;
 	public GameObject popupregion;
 	// Use this for initialization
 	void Start () {
@@ -14,9 +14,9 @@ public class IncreasePopularityorMoneyScript : MonoBehaviour {
 
 	// Update is called once per frame
 	void Update () {
-		
+
 	}
-	
+
 	void increasePopandMoney() {
 		Debug.Log ("clickyyyy");
 		popupregion.SendMessage ("IncreasePopularity", amounttoIncreasePop);
